JCR: RepoService refactored (prepared for other storages than segment tar)
