+++ gradle/libs.versions.toml
@@ -42,8 +42,7 @@ org-apache-directory-server-apacheds-server-integ = "1.5.7"
 org-apache-groovy = "4.0.22"
 org-apache-httpcomponents-client5-httpclient5 = "5.4.1"
 org-apache-httpcomponents-core5 = "5.3.1" # Needed by httpclient5
-org-apache-jackrabbit-oak-jcr = "1.66.0"
-org-apache-jackrabbit-oak-segment-tar = "1.66.0"
+org-apache-jackrabbit-oak = "1.74.0"
 org-apache-poi = "5.3.0"
 org-apache-poi-ooxml = "5.3.0"
 org-apache-wicket-myextensions = "10.3.0"
@@ -151,8 +150,8 @@ org-apache-commons-lang3 = { module = "org.apache.commons:commons-lang3", versio
 org-apache-commons-text = { module = "org.apache.commons:commons-text", version.ref = "org-apache-commons-text" }
 org-apache-groovy-all = { module = "org.apache.groovy:groovy-all", version.ref = "org-apache-groovy" }
 org-apache-httpcomponents-client5-httpclient5 = { module = "org.apache.httpcomponents.client5:httpclient5", version.ref = "org-apache-httpcomponents-client5-httpclient5" }
-org-apache-jackrabbit-oak-jcr = { module = "org.apache.jackrabbit:oak-jcr", version.ref = "org-apache-jackrabbit-oak-jcr" }
-org-apache-jackrabbit-oak-segment-tar = { module = "org.apache.jackrabbit:oak-segment-tar", version.ref = "org-apache-jackrabbit-oak-segment-tar" }
+org-apache-jackrabbit-oak-jcr = { module = "org.apache.jackrabbit:oak-jcr", version.ref = "org-apache-jackrabbit-oak" }
+org-apache-jackrabbit-oak-segment-tar = { module = "org.apache.jackrabbit:oak-segment-tar", version.ref = "org-apache-jackrabbit-oak" }
 org-apache-logging-log4j-api = { module = "org.apache.logging.log4j:log4j-api", version.ref = "org-apache-logging-log4j" }
 org-apache-logging-log4j-to-slf4j = { module = "org.apache.logging.log4j:log4j-to-slf4j", version.ref = "org-apache-logging-log4j" }
 org-apache-poi = { module = "org.apache.poi:poi", version.ref = "org-apache-poi" }
