+++ projectforge-jcr/src/main/kotlin/org/projectforge/jcr/SessionWrapper.kt
@@ -35,8 +35,8 @@ private val log = KotlinLogging.logger {}
 /**
  * Handles close and clean-up. Thin wrapper for [Session]
  */
-open class SessionWrapper(private val repoService: RepoService) {
-    val session: Session = repoService.repository.login(credentials)
+open class SessionWrapper(private val repoStore: OakStorage) {
+    val session: Session = repoStore.repository.login(credentials)
 
     val valueFactory: ValueFactory
         get() = session.valueFactory
@@ -57,7 +57,7 @@ open class SessionWrapper(private val repoService: RepoService) {
     }
 
     fun logout() {
-        repoService.fileStore?.flush()
+        repoStore.onLogout()
         session.save()
     }
